refactor(ui): move Step primitive into elements/ and tidy ProfileCard
- Promotes the Step compound primitive to packages/ui/src/components/
  ConfigureSSO/elements/Step.tsx — it lives alongside the other
  shared step-level primitives (Stepper, ProfileCard) instead of
  inside the steps/ folder.
- Updates all five step files to import Step from ../elements/Step.
- Drops ProfileCardBody and ProfileCardSection from
  elements/ProfileCard.tsx — those responsibilities now live on
  Step.Body and Step.Section, so the ProfileCard module is back to
  just Header and Footer chrome.
- ConfigureSSOSkeleton calls Stepper.Skeleton without an explicit
  totalSteps; the default of 4 is used.
